[Rename 🏷️] store => stores
